commit_templater: extract runtime objects from templater.rs
This should reduce future merge conflicts around use statements.
